Dont throw from Install-TestCertificate in test setup

+++ test/powershell/Modules/Microsoft.PowerShell.Security/CertificateProvider.Tests.ps1
@@ -82,8 +82,12 @@ Describe "Certificate Provider tests" -Tags "Feature" {
     BeforeAll{
         if($IsWindows)
         {
-            Install-TestCertificates
-            Push-Location Cert:\
+            if (-not (Install-TestCertificates) ) {
+                $SetupFailure = $true
+            } else {
+                $SetupFailure = $false
+                Push-Location Cert:\
+            }
         }
         else
         {
@@ -94,7 +98,7 @@ Describe "Certificate Provider tests" -Tags "Feature" {
     }
 
     AfterAll {
-        if($IsWindows)
+        if($IsWindows -and -not $SetupFailure)
         {
             Remove-TestCertificates
             Pop-Location
@@ -108,6 +112,9 @@ Describe "Certificate Provider tests" -Tags "Feature" {
     Context "Get-Item tests" {
         It "Should be able to get certifate by path: <path>" -TestCases $currentUserMyLocations {
             param([string] $path)
+
+            if ($SetupFailure) { Set-ItResult -Inconclusive -Because "Test certificates are not installed"}
+
             $expectedThumbprint = (Get-GoodCertificateObject).Thumbprint
             $leafPath = Join-Path -Path $path -ChildPath $expectedThumbprint
             $cert = (Get-Item -LiteralPath $leafPath)
@@ -116,6 +123,9 @@ Describe "Certificate Provider tests" -Tags "Feature" {
         }
         It "Should be able to get DnsNameList of certifate by path: <path>" -TestCases $currentUserMyLocations {
             param([string] $path)
+
+            if ($SetupFailure) { Set-ItResult -Inconclusive -Because "Test certificates are not installed"}
+
             $expectedThumbprint = (Get-GoodCertificateObject).Thumbprint
             $expectedName = (Get-GoodCertificateObject).DnsNameList
             $expectedEncodedName = (Get-GoodCertificateObject).DnsNameList
@@ -133,6 +143,9 @@ Describe "Certificate Provider tests" -Tags "Feature" {
         }
         it "Should be able to get EnhancedKeyUsageList of certifate by path: <path>" -TestCases $currentUserMyLocations {
             param([string] $path)
+
+            if ($SetupFailure) { Set-ItResult -Inconclusive -Because "Test certificates are not installed"}
+
             $expectedThumbprint = (Get-GoodCertificateObject).Thumbprint
             $expectedOid = (Get-GoodCertificateObject).EnhancedKeyUsageList[0].ObjectId
             $leafPath = Join-Path -Path $path -ChildPath $expectedThumbprint
@@ -144,6 +157,9 @@ Describe "Certificate Provider tests" -Tags "Feature" {
             $cert.EnhancedKeyUsageList[0].ObjectId | Should -Be $expectedOid
         }
         It "Should filter to codesign certificates" {
+
+            if ($SetupFailure) { Set-ItResult -Inconclusive -Because "Test certificates are not installed"}
+
             $allCerts = Get-Item cert:\CurrentUser\My\*
             $codeSignCerts = Get-Item cert:\CurrentUser\My\* -CodeSigningCert
             $codeSignCerts | Should -Not -Be null
@@ -152,6 +168,9 @@ Describe "Certificate Provider tests" -Tags "Feature" {
             $nonCodeSignCertCount | Should -Not -Be 0
         }
         It "Should be able to exclude by thumbprint" {
+
+            if ($SetupFailure) { Set-ItResult -Inconclusive -Because "Test certificates are not installed"}
+
             $allCerts = Get-Item cert:\CurrentUser\My\*
             $testThumbprint = (Get-GoodCertificateObject).Thumbprint
             $allCertsExceptOne = (Get-Item "cert:\currentuser\my\*" -Exclude $testThumbprint)
@@ -166,6 +185,9 @@ Describe "Certificate Provider tests" -Tags "Feature" {
             $cert = Get-GoodServerCertificateObject
         }
         it "Should filter to codesign certificates" {
+
+            if ($SetupFailure) { Set-ItResult -Inconclusive -Because "Test certificates are not installed"}
+
             $allCerts = get-ChildItem cert:\CurrentUser\My
             $codeSignCerts = get-ChildItem cert:\CurrentUser\My -CodeSigningCert
             $codeSignCerts | Should -Not -Be null
@@ -174,6 +196,9 @@ Describe "Certificate Provider tests" -Tags "Feature" {
             $nonCodeSignCertCount | Should -Not -Be 0
         }
         it "Should filter to ExpiringInDays certificates" {
+
+            if ($SetupFailure) { Set-ItResult -Inconclusive -Because "Test certificates are not installed"}
+
             $thumbprint = $cert.Thumbprint
             $NotAfter = $cert.NotAfter
             $before = ($NotAfter.AddDays(-1) - (Get-Date)).Days
@@ -186,6 +211,9 @@ Describe "Certificate Provider tests" -Tags "Feature" {
             $afterCerts.Thumbprint | Should -BeExactly $thumbprint
         }
         it "Should filter to DocumentEncryptionCert certificates" {
+
+            if ($SetupFailure) { Set-ItResult -Inconclusive -Because "Test certificates are not installed"}
+
             $thumbprint = $cert.Thumbprint
             $certs = Get-ChildItem cert:\CurrentUser\My\$thumbprint -DocumentEncryptionCert
 
@@ -199,6 +227,8 @@ Describe "Certificate Provider tests" -Tags "Feature" {
         ) {
             param($name, $searchName, $count, $thumbprint)
 
+            if ($SetupFailure) { Set-ItResult -Inconclusive -Because "Test certificates are not installed"}
+
             $certs = Get-ChildItem cert:\CurrentUser\My\$thumbprint -DNSName $searchName
 
             $certs.Count | Should -Be $count
@@ -206,6 +236,9 @@ Describe "Certificate Provider tests" -Tags "Feature" {
 
         }
         it "Should filter to SSLServerAuthentication certificates" {
+
+            if ($SetupFailure) { Set-ItResult -Inconclusive -Because "Test certificates are not installed"}
+
             $thumbprint = $cert.Thumbprint
 
             $certs = Get-ChildItem cert:\CurrentUser\My\$thumbprint -SSLServerAuthentication
@@ -220,6 +253,8 @@ Describe "Certificate Provider tests" -Tags "Feature" {
             @{ Name = "all patterns should be passed - negative test"; EKU = "*QWERTY*","*encryp*";                  Count = 0; Thumbprint = $null }
         ) {
             param($name, $ekuSearch, $count, $thumbprint)
+
+            if ($SetupFailure) { Set-ItResult -Inconclusive -Because "Test certificates are not installed"}
 
             $certs = Get-ChildItem cert:\CurrentUser\My\$thumbprint -EKU $ekuSearch
